Add Support for Album Cover Art Files (#176)
* add support for copying album art

* rewrite album art embedding and copying

This pulls all settings related to the album art into the alternatives
config, allowing to set up different things per external collection.

Also generally unifies the handling between linking / copy / converting and
embedding art.

* add album_art_* options to readme

* refactor album art action matching

moved the query matching inline to update_art as suggested and
refactored the matched_album_action function to remove some duplication
for the album art source and dest path logic

* fix pyright errors

* add testcase for no update of album art

also changed around the dest definition a bit as it was kinda eh to call
Path on that at every location.

* rename assert_has_artwork to assert_same_file_content

* refactored touch_art function

move at as a top-level function to the helpers and renamed the
arguments to make more sense

* add art_filename note to readme

* Simplify test_copy_art

* simplify album art tests

* refactor update_art to inline all conditions

* add changelog for album art files support

* ignore ResourceWarning for beets.util.artresizer

* apply suggested readme changes

Co-authored-by: Thomas Scholtes <geigerzaehler@axiom.fm>

* apply suggestions to simplify

Co-authored-by: Thomas Scholtes <geigerzaehler@axiom.fm>

* remove explicit en-/decoding to utf8

we want to default to the default codec for the filesystem

* fix update_art print output to full path

* assert that album art symlinks to correct file

* add back touch_art in album art tests

* revert using os.fsdecode for bytes -> Path conversion

---------

Co-authored-by: Thomas Scholtes <geigerzaehler@axiom.fm>